+++ src/tutorials.ts
@@ -1,10 +1,10 @@
-import helloWorld from "./tutorials/gno.land/gbe/basics/01-hello/index";
-import firstApp from "./tutorials/gno.land/gbe/basics/02-count/index";
-import primitives from "./tutorials/gno.land/gbe/basics/03-primitives/index";
-import variables from "./tutorials/gno.land/gbe/basics/04-variables/index";
-import constants from "./tutorials/gno.land/gbe/basics/05-constants/index";
-import conditions from "./tutorials/gno.land/gbe/basics/06-conditions/index";
-import loops from "./tutorials/gno.land/gbe/basics/07-loops/index";
+import helloWorld from "./tutorials/gno.land/gbe/getting-started/01-hello/index";
+import firstApp from "./tutorials/gno.land/gbe/getting-started/02-count/index";
+import primitives from "./tutorials/gno.land/gbe/getting-started/03-primitives/index";
+import variables from "./tutorials/gno.land/gbe/getting-started/04-variables/index";
+import constants from "./tutorials/gno.land/gbe/getting-started/05-constants/index";
+import conditions from "./tutorials/gno.land/gbe/getting-started/06-conditions/index";
+import loops from "./tutorials/gno.land/gbe/getting-started/07-loops/index";
 import helloWorld from "./tutorials/gno.land/gbe/new/01-hello/index";
 
 const tutorials = [
