+++ src/tutorials/gno.land/gbe/getting-started/01-hello/hello_test.gno
