+++ src/tutorials/gno.land/gbe/getting-started/02-count/index.ts
