+++ src/tutorials/gno.land/gbe/getting-started/03-primitives/index.ts
