+++ src/tutorials/gno.land/gbe/getting-started/03-primitives/primitives_test.gno
