+++ src/tutorials/gno.land/gbe/getting-started/04-variables/index.ts
