+++ src/tutorials/gno.land/gbe/getting-started/05-constants/constants.gno
