+++ src/tutorials/gno.land/gbe/getting-started/06-conditions/conditions_test.gno
