+++ src/tutorials/gno.land/gbe/getting-started/06-conditions/index.ts
