+++ src/tutorials/gno.land/gbe/getting-started/07-loops/index.ts
