+++ src/tutorials/gno.land/gbe/getting-started/07-loops/loops_test.gno
